app/requirements.txt
aiogram==3.13.1
python-dotenv==1.0.0
pytz==2025.2
